refactored the crate `sp_fragnova`
